feat: Implement worker assignment to map sources, refactor source entities to use double values, and add potential resource calculation.

+++ app/src/main/java/com/example/villageevo/repository/MapRepository.kt
@@ -28,7 +28,7 @@ class MapRepository(private val context: Context) {
                     it["id"]!!.toInt(),
                     it["id_map"]!!.toInt(),
                     it["name"]!!,
-                    it["value"]!!.toInt(),
+                    it["value"]!!.toDouble(),
                     it["x"]!!.toInt(),
                     it["y"]!!.toInt()
                 )
